// src/lib/types.ts
export type Craft = {
  id: number;
  name: string;
  state: string;
  district: string;
  category: string;
  material: string;
  technique: string;
  gi: boolean;
  artisan: string;
  language: string[];
  summary: string;
  history: string;
  authenticity: string;
  image: string;
};